Revert "add Long support"
This reverts commit dfb5f8bdf1ff73c96f583d2daa520cee3659395d.

+++ src/hmi/flipper/behaviourselection/template/value/AbstractValue.java
@@ -159,7 +159,7 @@ public class AbstractValue
 
         /* Depending on the type of the value, create a new Value */
         if( isInt && !isDouble && !isString ) {
-            return new Value( Long.parseLong(str) );
+            return new Value( Integer.parseInt(str) );
         } else if( isDouble && !isInt && !isString ) {
             return new Value( Double.parseDouble(str) );
         } else if( isString && !isInt && !isDouble ) {
